projects/atelier/spike-001/requirements.txt
anthropic==0.40.0
requests==2.32.3
google-search-results==2.4.2
Pillow==11.1.0
PyYAML==6.0.2
pytest==8.3.4
pytest-mock==3.14.0
click==8.1.8
